ResearchPaperNetworkV2/requirements.txt
Streamlit
PyMuPDF
nltk
google-generativeai